Ah feat 1402 06 08 (#33)
* feat(dashboard):environment change

clear baseUrl from environment

* feat(dashboard):Remove HomModule

delete extra Module from app

* add contactUs component & Fix Ux

* feat(dashboard):remmeberMe checkBox

modified remmeberMe checkBox

* feat(Back):fix RememberMe

feat(Back):fix RememberMe

* feat(dashboard):fix footer & styles

feat(dashboard):fix footer & styles

* feat(dashboard):Setup lazayLoading-1

* feat(Back):Add rememberMe Api

Add rememberMe Api and Services

* feat(dashboard):Add LoadingSpinner

Add LoadingSpinner

* feat(Back):increase Device Name char

feat(Back):increase Device Name char

* feat(dashboard):fix navbar bug

feat(dashboard):fix navbar bug in collapse and show

* feat(dashboard):fix extra call get request

* Revert "feat(Back):Add rememberMe Api"

This reverts commit 79782e6671a33bcf28cbafe1c07dfafe7e22a5bc.

* feat(dashboard):fix lazyLoading Bug

* feat(dashboard):add darkMode_1

* feat(dashboard):setup dark-mode-2

* feat(dashboard):setup sweetAlert

* feat(dashboard):responsive addButton

* feat(dashboard):responsive ContactUs

* feat(dashboard):BugFix-ContactUs alignment

* feat(dashboard):darkMode-edit3

* feat(dashboard):darkMode-e4

* feat(dashboard):logo position-1

* feat(dashboard):logo position fix-1

* feat(dashboard):about performance of dark mode

* feat(dashboard):about performance of dark mode2

* feat(dashboard):darkMod-4

* feat(dashboard):about performance of dark mode-e5

* feat(dashboard): fix responsive & darkMode

* feat(dashboard): fix navbar color in darkMode

* feat(Back): remove RememberMe from DB & code

* feat(dashboard):Fix automatic scroll Down

* feat(dashboard):Start Pwa Service-Worker

e1

* feat(dashboard):Upgrade to Angular 16 -1

* feat(dashboard):Upgrade to Angular 16 -2

* feat(dashboard):Upgrade to Angular 16 -3

* feat(dashboard):Upgrade to Angular 16 -4

* feat(dashboard):Upgrade to Angular 16 -5

* feat(dashboard):Upgrade to Angular 16 -6

* Revert "feat(dashboard):Upgrade to Angular 16 -6"

This reverts commit 96e527def0e1d7d3da610095097d6026a65a5764.

* Revert "feat(dashboard):Upgrade to Angular 16 -5"

This reverts commit f5b9538f0bbb06f1639a2d52351af7b94e272d0d.

* Revert "feat(dashboard):Upgrade to Angular 16 -4"

This reverts commit 85259a8a9cb17f6aa22a8238ea208307f606ff93.

* Revert "feat(dashboard):Upgrade to Angular 16 -3"

This reverts commit f3db105ffb20e6c86bb930dedf3a61a39c1c0cfd.

* 15

* 15

* Revert "feat(dashboard):Upgrade to Angular 16 -2"

This reverts commit 54550dee1c7e285bc4226f544151e13989c87567.

* Revert "feat(dashboard):Upgrade to Angular 16 -1"

This reverts commit 9550efab922f54e6438a8bb2981673d61829412c.

* feat(dashboard):Before Pwa Service-Worker

* feat(dashboard):Final PWA

* fixBug(dashboard):notFoundPage

* feat(dashboard):UX improve

* feat(dashboard):Fix darkMode pub & delete Console.log

* feat(dashboard):add sweetAlert

* fix(dashbaord): fix signUp message for sweetAlert

* feat(Back): Setup send Url to gmail

* feat(dashboard): New component changePassword

* back:emailservice

* back: fix email service

* feat(Back):setup Rest Password

* feat(dashboard):setup reset password

* bugFix(Back):fix expireDate in reset password

* feat(dashboard):send new pass to back

* feat(dashboard):fix Close resetPassword modal

* feat(Back):setup change password

* feat(dashboard) : setup change password

* fixBug(dashboard):fix display changePassword issue

* fixBug(dashboard):Fix display footer width issue

* fixBug(dashboard):changePassword

after chang Password, display succes
after reset Password, display succes

ux improve when mobile mode

* fix(Back):accountController changpassword

* feat(dashboard):increase speed of animation login

* fix(dashboard):delete extra console.log

* fix(Back):accountController

* fix(dashboard):fix color of signOut button

* fix(dashboard):fix navbar padding & show/hidden

* fix(dashboard):fix bug navbar resize to p-4

* fix(Back):appSetting.json

change address forgetbaseUrl

change password email sender

* feat dashboard

* feat(dashboard):fix publish issue , message SignUp

* feat(back):esmaili

* feat(back):complete reset password using token By Davood

* Fix(back) :  forgot password

* fixBug(back) : Bearer token forgot password Davood

* fix(back):forgot password token

* fixbug(backend) : forgot password token by davood

* feat(dashobaord):resetPasswordService

cahnge token

* fixbug(backeng) : reset password token

* feat (dashbaord): add spinner to forgotPass -1

* fix(dashboard): fix UI forgotDialog and Spinner

* fix(dashboard):fix spinnerLoading

* fix(dashboard):footer

* dashboard

---------

+++ src/Presentation/Presentation.DashBoard/src/app/clipBoard/clipBoard-item/clipBoard-item.component.ts
@@ -6,9 +6,7 @@ import {
   ElementRef,
   AfterViewInit,
 } from "@angular/core";
-import * as bootstrap from "bootstrap";
 import { fromEvent } from "rxjs";
-import { debounceTime } from "rxjs/operators";
 
 import { ColorUsedService } from "../../shared/services/color-used.service";
 import { MobileViewService } from "src/app/shared/services/mobile-view.service";
@@ -17,6 +15,7 @@ import { AddOrEditClipboardComponent } from "src/app/shared/modals/addOrEditClip
 import { MatDialog } from "@angular/material/dialog";
 import { DataSharingService } from "src/app/shared/services/data-sharing.service";
 import { ClipBoardService } from "src/app/shared/services/clipBoard.service";
+import Swal from "sweetalert2";
 
 @Component({
   selector: "app-clipBoard-item",
@@ -49,6 +48,7 @@ export class ClipBoardItemComponent implements OnInit, AfterViewInit {
   orange: string = this.colorUsedService.orange;
   blue: string = this.colorUsedService.blue;
   green: string = this.colorUsedService.green;
+  gray: string = this.colorUsedService.gray;
 
   isBigContent: boolean = false;
   isEditDialogOpen: boolean = false;
@@ -122,8 +122,9 @@ export class ClipBoardItemComponent implements OnInit, AfterViewInit {
   }
 
   openEditClipBoardDialog() {
-    console.log("open dialog Edit");
+    // console.log("open dialog Edit");
     this.editDialog.open(AddOrEditClipboardComponent, {
+      backdropClass: "true",
       data: {
         content: this.clipBoard.content,
         id: this.clipBoard.id,
@@ -139,18 +140,39 @@ export class ClipBoardItemComponent implements OnInit, AfterViewInit {
   // Delete Clipboard
   onDeleteClipBoard(event: Event) {
     if (event != null) {
-      console.log(this.clipBoard.id);
-      this.clipBoardService.DeleteClipBoard(this.clipBoard.id).subscribe({
-        // handle successful sign-up response
-        next: (response) => {
-          console.log(response),
-            // Reload the page after adding a new clipboard item
-
-            window.location.reload();
-        },
+      Swal.fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+        iconColor: this.orange,
+        confirmButtonColor: this.violet,
+        cancelButtonColor: this.gray,
+        confirmButtonText: "Yes, delete it!",
+      }).then((result) => {
+        if (result.isConfirmed) {
+          this.clipBoardService.DeleteClipBoard(this.clipBoard.id).subscribe({
+            // handle successful sign-up response
+            next: () => {
+              Swal.fire(
+                "Deleted!",
+                "Your file has been deleted.",
+                "success"
+              ).finally(() => window.location.reload());
+            },
+            // handle error
+            error: (errMes) => {
+              console.error(errMes),
+                Swal.fire({
+                  title: "Error!",
+                  text: errMes.error.detail,
+                  icon: "error",
+                  confirmButtonColor: this.violet,
+                });
+            },
+          });
+        }
       });
-    } else {
-      alert("خطا");
     }
   }
 }
